@@ -29,8 +29,8 @@ namespace GerenciadorDeJogos.Api.Controller
         }
 
         [HttpGet]
-        [Route("{id:Guid}/amigo")]
-        public async Task<ActionResult<AmigoResult>> ObterPorId(Guid id)
+        [Route("{id:int}/amigo")]
+        public async Task<ActionResult<AmigoResult>> ObterPorId(int id)
         {
             var amigo = await _amigoServico.BuscarPorIdAsync(id).ConfigureAwait(false);
             return Ok(amigo);
@@ -60,7 +60,7 @@ namespace GerenciadorDeJogos.Api.Controller
         }
 
         [HttpDelete]
-        public async Task<ActionResult<bool>> Deletar(Guid id)
+        public async Task<ActionResult<bool>> Deletar(int id)
         {
             var result = await (_amigoServico.ExcluirAsync(id)).ConfigureAwait(false);
             return Ok(result);
