@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace GerenciadorDeJogos.Application.Models.Request
+﻿namespace GerenciadorDeJogos.Application.Models.Request
 {
-   public class ModelBase
+    public class ModelBase
     {
-        public Guid Id { get; set; }
+        public int Id { get; set; }
     }
 }
