@@ -4,9 +4,9 @@ namespace GerenciadorDeJogos.Application.Models.Request
 {
     public class ItensDevolvidosRequest 
     {
-        public Guid EmprestimoId { get; set; }
+        public int EmprestimoId { get; set; }
 
-        public Guid JogoId { get; set; }
+        public int JogoId { get; set; }
 
         public bool? Devolvido { get; set; }
 
