@@ -4,8 +4,8 @@ namespace GerenciadorDeJogos.Application.Models.Request
 {
     public class ItensEmprestadosRequest:ModelBase
     {
-        public Guid JogoId { get; set; }
+        public int JogoId { get; set; }
 
-        public Guid EmprestimoId { get; set; }
+        public int EmprestimoId { get; set; }
     }
 }
