@@ -5,6 +5,6 @@ namespace GerenciadorDeJogos.Application.Models.Request
     public class JogoRequest: ModelBase
     {
         public string Nome { get; set; }
-        public Guid ProprietarioId { get; set; }
+        public int ProprietarioId { get; set; }
     }
 }
