@@ -5,11 +5,11 @@ namespace GerenciadorDeJogos.Application.Models.Result
 {
     public class ItensEmprestadosResult:ModelBase
     {
-        public Guid JogoId { get; set; }
+        public int JogoId { get; set; }
 
         public string NomeJogo { get; set; }
 
-        public Guid EmprestimoId { get; set; }
+        public int EmprestimoId { get; set; }
 
         public DateTime? DataDevolucao { get; set; }
 
