@@ -6,7 +6,7 @@ namespace GerenciadorDeJogos.Application.Models.Result
     public class JogoResult: ModelBase
     {
         public string Nome { get; set; }
-        public Guid ProprietarioId { get; set; }
+        public int ProprietarioId { get; set; }
         public string NomeProprietario { get; set; }
     }
 }
