@@ -1,12 +1,13 @@
 ﻿using GerenciadorDeJogos.Application.Models.Request;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace GerenciadorDeJogos.Application.Models.Result
 {
     public class UsuarioResult:ModelBase
     {
-        public int MyProperty { get; set; }
+        public string Nome { get; set; }
+
+        public string Login { get; set; }
+
+        public string Senha { get; set; }
     }
 }
