@@ -10,11 +10,11 @@ namespace GerenciadorDeJogos.Application.Repositorios
     public interface IRepositorioBase<T> where T : Entidade
     {
         T Inserir(T item);
-        T BuscarPorId(Guid id, params Expression<Func<T, object>>[] incluindoPropriedades);
+        T BuscarPorId(int id, params Expression<Func<T, object>>[] incluindoPropriedades);
         IQueryable<T> ListarTodos();
         T Atualizar(T item);
-        void Excluir(Guid id);
-        bool Existe(Guid id);
+        void Excluir(int id);
+        bool Existe(int id);
         ListaPaginavel<T> Listar(int indiceDaPagina, int tamanhoDaPagina, Expression<Func<T,IKey>> chaveSeletora, TipoDeOrdenacao orderBy);
 
         ListaPaginavel<T> Listar(int indiceDaPagina, int tamanhoDaPagina, Expression<Func<T,IKey>> chaveSeletora,
