@@ -4,9 +4,9 @@ namespace GerenciadorDeJogos.Domain.Entidades.Base
 {
     public class PesquisaEmprestimo: Pesquisa
     {
-        public Guid? AmigoId { get; set; }
+        public int? AmigoId { get; set; }
 
-        public Guid? JogoId { get; set; }
+        public int? JogoId { get; set; }
 
         public DateTime? DataEmprestimo { get; set; }
     }
